+++ 6. Backend Beginner/tickitz-moviez/src/controllers/movies.js
@@ -4,8 +4,12 @@ const response = require('../helpers/response')
 
 ctrl.addMovie = async (req, res) => {
   try {
-    const data = await model.addMovie({ ...req.body, ...req.file })
-    return response(res, 201, data)
+    if (req.decode.role === 'admin') {
+      const data = await model.addMovie({ ...req.body, ...req.file })
+      return response(res, 201, data)
+    } else {
+      return response(res, 401, 'Anda tidak memiliki akses', true)
+    }
   } catch (error) {
     return response(res, 500, 'Terjadi kesalahan', true)
   }
@@ -13,7 +17,12 @@ ctrl.addMovie = async (req, res) => {
 
 ctrl.getAllMovies = async (req, res) => {
   try {
-    const data = await model.getAllMovies()
+    const pagination = {
+      page: req.query.page || 1,
+      limit: req.query.limit || 5,
+      order: req.query.order
+    }
+    const data = await model.getAllMovies(pagination)
     return response(res, 200, data)
   } catch (error) {
     return response(res, 500, 'Terjadi kesalahan', true)
@@ -49,12 +58,16 @@ ctrl.searchMovie = async (req, res) => {
 
 ctrl.updateMovie = async (req, res) => {
   try {
-    const data = await model.updateMovie({
-      ...req.params,
-      ...req.body,
-      ...req.file
-    })
-    return response(res, 200, data)
+    if (req.decode.role === 'admin') {
+      const data = await model.updateMovie({
+        ...req.params,
+        ...req.body,
+        ...req.file
+      })
+      return response(res, 200, data)
+    } else {
+      return response(res, 401, 'Anda tidak memiliki akses', true)
+    }
   } catch (error) {
     return response(res, 500, 'Terjadi kesalahan', true)
   }
@@ -62,8 +75,12 @@ ctrl.updateMovie = async (req, res) => {
 
 ctrl.deleteMovie = async (req, res) => {
   try {
-    const data = await model.deleteMovie(req.params)
-    return response(res, 200, data)
+    if (req.decode.role === 'admin') {
+      const data = await model.deleteMovie(req.params)
+      return response(res, 200, data)
+    } else {
+      return response(res, 401, 'Anda tidak memiliki akses', true)
+    }
   } catch (error) {
     return response(res, 500, 'Terjadi kesalahan', true)
   }
