+++ 6. Backend Beginner/tickitz-moviez/src/controllers/schedules.js
@@ -4,8 +4,12 @@ const response = require('../helpers/response')
 
 ctrl.addSchedule = async (req, res) => {
   try {
-    const data = await model.addSchedule(req.body)
-    return response(res, 201, data)
+    if (req.decode.role === 'admin') {
+      const data = await model.addSchedule(req.body)
+      return response(res, 201, data)
+    } else {
+      return response(res, 401, 'Anda tidak memiliki akses', true)
+    }
   } catch (error) {
     return response(res, 500, 'Terjadi kesalahan', true)
   }
@@ -31,8 +35,12 @@ ctrl.getScheduleById = async (req, res) => {
 
 ctrl.updateSchedule = async (req, res) => {
   try {
-    const data = await model.updateSchedule({ ...req.params, ...req.body })
-    return response(res, 200, data)
+    if (req.decode.role === 'admin') {
+      const data = await model.updateSchedule({ ...req.params, ...req.body })
+      return response(res, 200, data)
+    } else {
+      return response(res, 401, 'Anda tidak memiliki akses', true)
+    }
   } catch (error) {
     return response(res, 500, 'Terjadi kesalahan', true)
   }
@@ -40,8 +48,12 @@ ctrl.updateSchedule = async (req, res) => {
 
 ctrl.deleteSchedule = async (req, res) => {
   try {
-    const data = await model.deleteSchedule(req.params)
-    return response(res, 200, data)
+    if (req.decode.role === 'admin') {
+      const data = await model.deleteSchedule(req.params)
+      return response(res, 200, data)
+    } else {
+      return response(res, 401, 'Anda tidak memiliki akses', true)
+    }
   } catch (error) {
     return response(res, 500, 'Terjadi kesalahan', true)
   }
